backend/requirements.txt
# Clinical GraphRAG Pro — Backend Dependencies

# ── Web Framework ────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.2
sse-starlette==2.1.3

# ── Settings ─────────────────────────────────────────────
pydantic-settings==2.7.1

# ── Database ─────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# ── Redis ────────────────────────────────────────────────
redis[hiredis]==5.2.1

# ── AI / ML ──────────────────────────────────────────────
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
rank-bm25>=0.2.2
numpy==1.26.4

# ── HTTP Client ──────────────────────────────────────────
httpx==0.28.1

# ── PDF Processing ───────────────────────────────────────
PyPDF2==3.0.1

# ── Utilities ────────────────────────────────────────────
python-dotenv==1.0.1
PyJWT>=2.8.0

# ── Testing ──────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
